requirements.txt
fastapi[standard]
uvicorn
fastapi-analytics
slowapi
tokencost
pytest
pytest-asyncio
click>=8.0.0
